+++ src/frontend/js/widgets/Dashboard/components/TeacherOrganizationDashboardSidebar/index.tsx
@@ -1,12 +1,18 @@
-import { defineMessages, useIntl } from 'react-intl';
+import { FormattedMessage, defineMessages, useIntl } from 'react-intl';
 import { generatePath, useParams } from 'react-router-dom';
 import { useMemo } from 'react';
 import { TeacherDashboardPaths } from 'widgets/Dashboard/utils/teacherRouteMessages';
-import { DashboardSidebar } from 'widgets/Dashboard/components/DashboardSidebar';
+import {
+  DashboardAvatarPositionEnum,
+  DashboardSidebar,
+} from 'widgets/Dashboard/components/DashboardSidebar';
 import {
   getDashboardRouteLabel,
   getDashboardRoutePath,
 } from 'widgets/Dashboard/utils/dashboardRoutes';
+import { useOrganization } from 'hooks/useOrganizations';
+import { Spinner } from 'components/Spinner';
+import { DashboardAvatar, DashboardAvatarVariantEnum } from '../DashboardAvatar';
 
 const messages = defineMessages({
   subHeader: {
@@ -14,13 +20,22 @@ const messages = defineMessages({
     description: 'Sub title of the organization dashboard sidebar',
     defaultMessage: 'You are on the organization dashboard',
   },
+  loading: {
+    defaultMessage: 'Loading organization...',
+    description: 'Message displayed while loading an organization',
+    id: 'components.TeacherOrganizationDashboardSidebar.loading',
+  },
 });
 
 export const TeacherOrganizationDashboardSidebar = () => {
   const intl = useIntl();
   const getRoutePath = getDashboardRoutePath(intl);
   const getRouteLabel = getDashboardRouteLabel(intl);
-  const params = useParams();
+  const { organizationId } = useParams<{ organizationId: string }>();
+  const {
+    item: organization,
+    states: { fetching },
+  } = useOrganization(organizationId);
 
   const links = useMemo(
     () =>
@@ -33,18 +48,36 @@ export const TeacherOrganizationDashboardSidebar = () => {
           getRoutePath(path, {
             organizationId: ':organizationId',
           }),
-          params,
+          { organizationId },
         ),
         label: getRouteLabel(path),
       })),
     [],
   );
 
+  if (fetching) {
+    return (
+      <Spinner aria-labelledby="loading-courses-data">
+        <span id="loading-courses-data">
+          <FormattedMessage {...messages.loading} />
+        </span>
+      </Spinner>
+    );
+  }
+
   return (
     <DashboardSidebar
       menuLinks={links}
-      header="Dummy Organization"
+      header={organization.title}
       subHeader={intl.formatMessage(messages.subHeader)}
+      avatarPosition={DashboardAvatarPositionEnum.THREE_QUARTER}
+      avatar={
+        <DashboardAvatar
+          title={organization.title}
+          variant={DashboardAvatarVariantEnum.SQUARE}
+          imageUrl={organization.logo.src}
+        />
+      }
     />
   );
 };